Create Day 2: Operators
Day 2 Operators Solution is the part of Hackerrank 30 days of code, in this problem. Given the meal price (base cost of a meal), tip per cent (the percentage of the meal price being added as a tip), and tax per cent (the percentage of the meal price being added as tax) for a meal, find and print the meal’s total cost.

Note: Be sure to use precise values for your calculations, or you may end up with an incorrectly rounded result!

Input Format for Day 2 Operators
There are 3 lines of numeric input:
The first line has a double, mealCost(the cost of the meal before tax and tip).
The second line has an integer, tipPercent(the percentage of mealCost being added as a tip).
The third line has an integer, taxPercent(the percentage of mealCost being added as tax).

Output Format for Operators Hackerrank Solution
Print the total meal cost, where totalCost is the rounded integer result of the entire bill (mealCost with added tax and tip).

Logic
Here we go for day 2 Hackerrank solutions in C. We can solve this problem by taking a simple example. Suppose you are in the cafeteria and you order meal after you finish your meal you have to pay some amount of money also assume you are also interested to pay a Tip(extra amount of money almost 15 – 20 percent of your bill).

You also have to pay Tax’s so what will be the total amount you paid in the cafeteria. This is a problem for that we will take hacker rank example and solve the problem step by step.

day 2 explanation
Explanation
As we are taking hacker rank example so Meal cost is 12 dollar (according to hacker rank) Tip per cent 20 and tax per cent is 8 so for according to above input our program perform following steps.

Step:1 Tip = Mealcost * tip percent / 100

Tip = 12 * 20 / 100 = 2.4

Step 2:- Tax = mealcost * tax percent / 100

Tax = 12 * 8 / 100 = 0.96

Step 3:- Total = Mealcost + Tip +Tax.

Total = 12 + 2.4 +0.96 = 15.36

Total = round(Total) = 15.36 So final Bill is 15 Dollar paid by customer in cafeteria. Below is the Day 2 Operators Solution in C Language