Create Day 3: Intro to Conditional Statements
Hackerrank Day 3 Intro to Conditional Statements Solution in C, C++, and Java programming language. In this programming problem, we are going to learn about the If-Else statements, We have to use Even-Odd logic, and Greater and Less Number logic to solve this Programming Problem of Hackerrank Website.

Problem Statement:- According to Hackerrank, you should follow these four given conditions. Hackerrank Intro to Conditional Statements Solution in C.

If ‘n‘ is odd, print Weird.
If ‘n‘ is even and in the inclusive range of 2 to 5, print Not Weird.
‘n‘ is even and in the inclusive range of 6 to 20, print Weird.
If ‘n‘ is even and greater than 20, print Not Weird.
Input Output for Intro to Conditional Statements
Intro to Conditional Statements Solution Explained
Let’s start, so we have a number n and we need to write a program that follows the above four condition as we can see that our first condition is if n is odd that program will print number “Weird“.

So for this first condition, we divide a number by 2 if a number is divisible by 2 then the number is even and if the number is not divisible by 2 then it will print the number is “Weird‘.

Now come to the second condition if the number is even and range of 2 to 5 the program will print the number “Not Weird”, for that again each number we divide by 2 if a number is even and the number is between range(2 to 5) the program will print Number is “Not Weird“.

Again comes to the third condition if the number is between 6 to 20 then the program will print the number is ” Weird”. same as the second condition. Again check for the fourth condition if the number is greater than 20 then the program will print the number is “Weird“.